requirements.txt
Django==5.2.1
djangorestframework==3.15.1
drf-yasg==1.21.7
social-auth-app-django==5.4.1
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
